Improved Snackbar validation
